requirements.txt
# SQL Query Buddy - Requirements
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.10.0

# Vector Database
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Database Connectivity
sqlalchemy>=2.0.25
pymysql>=1.1.0
psycopg2-binary>=2.9.9
duckdb>=0.10.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3

# UI
gradio>=4.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
tiktoken>=0.5.2

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0

# Optional: For better SQL parsing
sqlparse>=0.4.4

